resources/python/app/requirements.txt
# Sidecar dependencies. Pinned to majors known to be CPU/MPS-compatible on
# darwin-arm64 with the python-build-standalone CPython 3.12 interpreter.
# Avoid CUDA-only torch wheels — pip will pick the macOS arm64 wheel.

# WhisperX bundles faster-whisper + alignment tooling.
whisperx>=3.1.1,<4.0.0
faster-whisper>=1.0.3,<2.0.0

# Diarization.
pyannote.audio>=3.1.0,<4.0.0

# Torch — pip resolves the macOS arm64 wheel; do not pin a CUDA build.
torch>=2.2,<3.0
torchaudio>=2.2,<3.0

# Numerics + IO helpers.
numpy>=1.26,<3.0
ffmpeg-python>=0.2.0
